move synthetizeData to api utils so we can share it
